+++ packages/varlet-ui/src/locale/fa-IR.d.ts
@@ -1,0 +1,5 @@
+import type { Pack } from '../../types'
+
+declare const faIR: Pack
+
+export default faIR