+++ packages/varlet-ui/src/switch/example/index.vue
@@ -9,9 +9,16 @@ const values = reactive({
   value2: true,
   value3: true,
   value4: true,
+  value5: true,
 })
 
-const { value, value1, value2, value3, value4 } = toRefs(values)
+const { value, value1, value2, value3, value4, value5 } = toRefs(values)
+
+function handleBeforeChange(value, change) {
+  setTimeout(() => {
+    change(value)
+  }, 500)
+}
 
 watchLang(use)
 onThemeChange()
@@ -37,5 +44,11 @@ onThemeChange()
   <app-type>{{ pack.loading }}</app-type>
   <var-switch :model-value="true" loading />
   <app-type>{{ pack.validateValue }}</app-type>
-  <var-switch v-model="value4" :rules="[(v) => v === true || pack.state]" />
+  <var-space :size="['3vmin', '9vmin']">
+    <var-switch v-model="value4" :rules="[(v) => v === true || pack.state]" />
+  </var-space>
+  <app-type>{{ pack.lazyChange }}</app-type>
+  <var-space :size="['3vmin', '9vmin']">
+    <var-switch v-model="value5" lazy-change @before-change="handleBeforeChange" />
+  </var-space>
 </template>