+++ packages/varlet-ui/src/switch/props.ts
@@ -1,6 +1,8 @@
 import { type PropType } from 'vue'
 import { defineListenerProp } from '../utils/components'
 
+export type ValidateTrigger = 'onChange' | 'onLazyChange'
+
 export const props = {
   modelValue: {
     default: false,
@@ -11,6 +13,7 @@ export const props = {
   inactiveValue: {
     default: false,
   },
+  lazyChange: Boolean,
   disabled: Boolean,
   readonly: Boolean,
   loading: Boolean,
@@ -18,12 +21,17 @@ export const props = {
   loadingColor: String,
   closeColor: String,
   size: [String, Number],
-  rules: Array as PropType<Array<(v: boolean) => any>>,
+  rules: Array as PropType<Array<(v: any) => any>>,
   ripple: {
     type: Boolean,
     default: true,
   },
+  validateTrigger: {
+    type: Array as PropType<Array<ValidateTrigger>>,
+    default: () => ['onChange', 'onLazyChange'],
+  },
   onClick: defineListenerProp<(event: Event) => void>(),
-  onChange: defineListenerProp<(value: boolean) => void>(),
-  'onUpdate:modelValue': defineListenerProp<(value: boolean) => void>(),
+  onBeforeChange: defineListenerProp<(value: any, change: (value: any) => void) => void>(),
+  onChange: defineListenerProp<(value: any) => void>(),
+  'onUpdate:modelValue': defineListenerProp<(value: any) => void>(),
 }