+++ packages/varlet-ui/types/indexBar.d.ts
@@ -26,6 +26,7 @@ export class IndexBar extends VarComponent {
 
   $slots: {
     default(): VNode[]
+    'anchor-name'(anchorName: string): VNode[]
   }
 
   scrollTo: (index: number | string, options?: IndexBarScrollToOptions) => void