+++ packages/varlet-ui/types/rate.d.ts
@@ -21,7 +21,7 @@ export interface RateProps extends BasicAttributes {
   readonly?: boolean
   clearable?: boolean
   ripple?: boolean
-  rules?: Array<(value: any) => any>
+  rules?: Array<(value: number) => any>
   onChange?: ListenerProp<(value: number) => void>
   'onUpdate:modelValue'?: ListenerProp<(value: number) => void>
 }